jobs/dbt_vivo/pyproject.toml
[project]
name = "dbt-vivo"
version = "0.1.0"
description = ""
authors = [
    {name = "Lasnami Abdelkarim",email = "81482604+LasnamiWalid16@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-bigquery (>=1.9.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "dbt_vivo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
